# server/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nothing to see here",
        "title": "An API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/posts": {
            "get": {
                "description": "Get all posts",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get all posts",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/posts/clean_positions": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Clean post positions",
                "tags": [
                    "posts"
                ],
                "summary": "Clean post positions",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/posts/make_post": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set a post as posted",
                "tags": [
                    "posts"
                ],
                "summary": "Set a post as posted",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "description": "Get a post",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a post",
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/posts/{id}/unpost": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set a post as unposted",
                "tags": [
                    "posts"
                ],
                "summary": "Set a post as unposted",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Post ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Please provide a valid api token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}